% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict.R
\name{svdgppredict}
\alias{svdgppredict}
\title{SVDpredict: Spatio-Temporal Prediction Using SVD-GP Model}
\usage{
svdgppredict(model, Dpred, computecov = FALSE)
}
\arguments{
\item{model}{A trained SVD-GP model obtained using the \code{svdgppmodel()} function.}

\item{Dpred}{A matrix of new spatial locations where predictions are to be made.}

\item{computecov}{Logical. If \code{TRUE}, computes the posterior covariance matrix for the predictions (default: \code{FALSE}).}
}
\value{
A list containing the following elements:
\item{Mn}{A matrix of predicted spatio-temporal values. Each row corresponds to a spatial prediction location, and each column represents a temporal point.}
\item{Kn}{The posterior covariance matrix for the predicted spatio-temporal values (only included if \code{computecov = TRUE}).}
\item{Coefpred}{Predictions of the coefficients for the \code{K} spatial components modeled by Gaussian Processes. Each column corresponds to one spatial component.}
\item{spcov}{A list of covariance matrices for each predicted spatial component. Each entry corresponds to one of the \code{K} components.}
\item{tempcov}{The temporal covariance matrix derived from the singular value decomposition. This matrix captures the temporal relationships between the modes.}
}
\description{
This function predicts spatio-temporal values using a trained SVD-GP model. It provides predictions for the spatio-temporal field, spatial components, temporal components, and associated covariance matrices.
}
\examples{
# Example usage:
# Test function
fxt <- function(x, t) { sin(2 * pi * t * (x - t)) }

# Design matrix
D <- matrix(seq(0, 1, length.out = 5), ncol = 1)

# Discretized time points
tt <- seq(0, 1, length.out = 15)

# Spatio-temporal observations matrix
FD <- apply(array(tt), 1, function(t) {
  apply(D, 1, function(x) { fxt(x, t) })
})

# Perform SVD and Gaussian Process modeling
model <- svdgppmodel(D, FD, formula = ~1, K = 4, spcovtype = "matern5_2",typekrig = "SK")
Dpred <- matrix(runif(20), ncol=1)  # New spatial locations
Yp <- svdgppredict(model, Dpred, computecov=TRUE)
print(Yp$Mn)
print(Yp$Kn)
}
